first attempt with metadata inline
