+++ static/js/apps/explore/success_result.tsx
@@ -192,6 +192,27 @@ export function SuccessResult(props: SuccessResultPropType): ReactElement {
                     maxBlock={maxBlock}
                   />
                 )}
+                {props.highlightPageMetadata && (
+                  <div
+                    className="related-data-explore"
+                    style={{
+                      margin: "20px 0",
+                      padding: "10px",
+                      backgroundColor: "#f9f9f9",
+                      borderRadius: "5px",
+                    }}
+                  >
+                    <p
+                      style={{
+                        fontSize: "16px",
+                        fontWeight: "bold",
+                        color: "#333",
+                      }}
+                    >
+                      Here is more related data to explore:
+                    </p>
+                  </div>
+                )}
                 <SubjectPageMainPane
                   id={PAGE_ID}
                   place={props.pageMetadata.place}
